smolagents-code-commenter/requirements.txt
smolagents>=0.1.0
openai>=1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
requests
pydantic>=2.10.2
python-dotenv>=1.0.1
httpx>=0.24.0
daytona-sdk>=0.6.0
api-client>=1.3.1
